improved evolutioner

+++ evolutioners/individual.py
@@ -4,6 +4,7 @@ import asyncio
 import os
 import json
 import time
+import pickle
 from agents import AnyAgent, TrainingProgress
 from environments import Environment
 
@@ -18,7 +19,7 @@ class Individual:
         self.history.append(progress.total_reward)
         return False
 
-    def life(self, dump_file: str = "") -> T.List[float]:
+    def life(self, dump_file: str = "", state_dump_file: str = "") -> T.List[float]:
         self.agent.add_progress_callback(self._on_progress)
         self.agent.tensor_board_log = False
         self.agent.save_progress = False
@@ -26,12 +27,16 @@ class Individual:
         self.env.close()
         if dump_file != "":
             json.dump(self.history, open(dump_file, "w"))
+        if state_dump_file != "":
+            pickle.dump(self.agent.get_state_dict(), open(state_dump_file, "wb"))
         return self.history
 
-    async def parallel_life(self, worker_id: int, timeout: int = 0) -> T.List[float]:
-        file_name = str(worker_id) + "_" + str(time.time()) + ".json"
+    async def parallel_life(self, worker_id: int, timeout: int = 0) -> T.Tuple[T.List[float], dict]:
+        now = str(time.time())
+        file_name = f"{worker_id}_result_{now}.json"
+        state_file_name = f"{worker_id}_state_dict_{now}.json"
 
-        process = multiprocessing.Process(target=self.life, args=(file_name,))
+        process = multiprocessing.Process(target=self.life, args=(file_name, state_file_name))
         process.start()
         start = time.time()
         while True:
@@ -39,11 +44,13 @@ class Individual:
                 process.kill()
                 raise TimeoutError("worker "+str(worker_id)+" timeout")
 
-            if os.path.isfile(file_name):
+            if os.path.isfile(file_name) and os.path.isfile(state_file_name):
                 await asyncio.sleep(1)
                 result = json.load(open(file_name))
+                state_dict = pickle.load(open(state_file_name, "rb"))
                 os.remove(file_name)
-                return result
+                os.remove(state_file_name)
+                return result, state_dict
             elif not process.is_alive():
                 raise RuntimeError("worker "+str(worker_id)+" died without returning a result")
